requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Bcrypt==1.0.1
Flask-Migrate==4.0.7
python-dateutil==2.8.2
Werkzeug==2.3.7
SQLAlchemy==2.0.36
Jinja2==3.1.4
WTForms==3.0.1
email-validator==2.1.0.post1
python-dotenv==1.0.0
gunicorn==21.2.0
